Calibration quality

Assesses the model's calibration quality via the following metrics:

NSE: Nash-Sutcliffe Efficiency

mNSE: Modified Nash-Sutcliffe Efficiency (j = 1)

RMSE: Root Mean Square Error

NRMSE: Normalised Root Mean Square Error (normalised by the standard deviation of the observations)

PBIAS: Percent Bias

RSR: Ratio of the RMSE to the standard deviation of the observations

GMRAE: Geometric Mean Relative Absolute Error

MdRAE: Median Relative Absolute Error

In addition, a scatter plot with the observed loads on the x- and the modelled loads on the y-axis is displayed and provides a visual impression of the model performance. Other elements of this plot are an identity line (solid) and plus/minus 30% deviation lines (dashed).

# S4 method for RPhosFate
calibrationQuality(x, substance, col)

Arguments

x
An S4 RPhosFate river catchment object.

substance
A character string specifying the substance to calculate.

col
A character string specifying the calibration data column with the respective substance loads.

Value

A named numeric vector containing the assessed metrics along with the in-channel retention.

See also