Demonstration project

Copies a demonstration project to an existing or a temporary directory.

The demonstration project data are a derivative of the

Geoland.at (digital elevation model),

AMA (field data; utilised dataset with the ID 35e36014-ec69-439b-8629-389f52ffaa92 was removed from Offene Daten Österreich),

BMLRT (channel data) and

GIP.at (road data)

data sets, used and licensed under (CC BY 4.0) by Gerold Hepp.

While the data represent a real catchment (HOAL), some of them are fictitious, but plausible. These are, among others, R- and C-factors, soil and related data, existence of subsurface drainage at road embankments as well as substance river loads.

Usage

demoProject(cs_dir = tempdir(TRUE))

Arguments

cs_dir

An optional character string specifying an existing directory.

Value

A character string containing the demonstration project root directory.

See also

Examples

demoProject()
#> Warning: A folder called "demoProject" already exists and is left as is.
#> [1] "/tmp/RtmpyttRnh/demoProject"