Subsequent run

Calls transport for the specified substance and optionally erosionPrerequisites, erosion, emission and transportPrerequisites beforehand.

Usage

# S4 method for class 'RPhosFate'
subsequentRun(
  x,
  substance = "PP",
  erosionPrerequisites = FALSE,
  erosion = FALSE,
  emission = FALSE,
  transportPrerequisites = FALSE
)

Arguments

x

An S4 RPhosFate river catchment object.

substance

A character string specifying the substance to calculate.

erosionPrerequisites

A logical scalar specifying if erosionPrerequisites is called.

erosion

A logical scalar specifying if erosion is called.

emission

A logical scalar specifying if emission is called. It is never called with substance = "SS" though.

transportPrerequisites

A logical scalar specifying if transportPrerequisites is called.

Value

An S4 RPhosFate river catchment object and side effects in the form of raster files.

See also

Examples

# \donttest{
# temporary demonstration project copy
cv_dir <- demoProject()
#> Warning: A folder called "demoProject" already exists and is left as is.
# load temporary demonstration project
x <- RPhosFate(
  cv_dir = cv_dir,
  ls_ini = TRUE
)
# presupposed method call
x <- firstRun(x, "SS")

x <- subsequentRun(x, "PP")# }