DEM Related Input

Clips, pre-processes and calculates or determines all input data related to the digital elevation model (DEM) in the broader sense: acc, acc_wtd, cha, dem, dir, rds, slp, and wsh.

Requires TauDEM 5.3.7 to be installed on your computer and will download the WhiteboxTools binary if needed.

DEMrelatedInput(cv_dir, cs_dem, cs_cha, sp_msk, sp_olp, sp_sds,
  cs_rds = NULL, cs_wgs = NULL, ns_brn = 50, is_adj = 1L,
  is_ths = 1L, ls_tmp = FALSE)

Arguments

cv_dir
A character vector specifying the desired project root directory (first position).

cs_dem
A character string specifying a path to a potentially large raster digital elevation model.

cs_cha
A character string specifying a path to a potentially large raster providing channels.

sp_msk
An sp::SpatialPolygonsDataFrame providing a somewhat oversized catchment mask used to clip the potentially large input rasters for further processing.

sp_olp
An sp::SpatialPointsDataFrame providing the desired catchment outlet.

sp_sds
An sp::SpatialPointsDataFrame providing channel sources.

cs_rds
An optional character string specifying a path to a potentially large raster providing roads.

cs_wgs
An optional character string specifying a path to a potentially large raster providing flow accumulation weights.

ns_brn
A numeric scalar specifying the stream burning step size in m.

is_adj
A numeric scalar specifying how many cells adjacent to channels shall be burnt.

is_ths
An integer scalar specifying the number of threads to use during computation.

ls_tmp
A logical scalar specifying if the temporary files created during computation shall be kept.

Value

A numeric matrix specifying the catchment outlet coordinates.

Details

This function applies the following (pre-processing) steps to ensure hydrologic consistency of the generated input data:

Stream burning and orientation of cells adjacent to channel cells approximately into the direction of channel cells (optional).

Depression breaching.

Tracing of downslope flowpaths from the provided channel sources.

In case no flow accumulation weights are provided, acc and acc_wtd are identical.

slp is calculated from the original, unprocessed DEM and represents D8 slopes.

References

Lindsay, J.B., 2016. Efficient hybrid breaching-filling sink removal methods for flow path enforcement in digital elevation models. Hydrological Processes 30, 846–857.

See also