Automatic model calibration

Automatically calibrates the model with the help of a combination of golden section search and successive parabolic interpolation.

# S4 method for RPhosFate
autoCalibrate(
  x,
  substance,
  col,
  interval,
  metric,
  tol = min(interval) * 0.1,
  parameter = NULL
)

Arguments

x

An S4 RPhosFate river catchment object.

substance

A character string specifying the substance to calculate.

col

A character string specifying the calibration data column with the respective substance river loads.

interval

A numeric vector specifying the end-points of the interval to be searched.

metric

A character string specifying the metric to optimise. See calibrationQuality for available metrics.

tol

A numeric scalar specifying the desired accuracy of the parameter used for optimisation (not the metric).

parameter

By default, SS are optimised utilising the overland deposition rate and all other substances are optimised utilising their respective enrichment ratio. This argument can be used to specify a dedicated parameter utilised for optimisation via a character string (overland ("ns_dep_ovl") or channel deposition rate ("ns_dep_cha")).

Value

An S4 RPhosFate river catchment object and side effects in the form of raster files.

See also

Examples

if (FALSE) {
# temporary demonstration project copy
cv_dir <- demoProject()
# load temporary demonstration project
x <- RPhosFate(
  cv_dir = cv_dir,
  ls_ini = TRUE
)
# presupposed method call
x <- firstRun(x, "SS")
x <- snapGauges(x)

x <- autoCalibrate(
  x,
  "SS",
  col = "SS_load",
  interval = c(10e-4, 20e-4),
  metric = "NSE"
)
}