First run

Calls erosionPrerequisites, erosion, emission, transportPrerequisites, transportCalcOrder and transport in the mentioned order. While transport is called for the specified substance only, emission is called for all substances whose top soil concentrations have been provided.

# S4 method for RPhosFate
firstRun(x, substance = "PP")

Arguments

x
An S4 RPhosFate river catchment object.

substance
A character string specifying the substance to calculate.

Value

An S4 RPhosFate river catchment object.

See also