Subsequent run

Calls transport for the specified substance and optionally erosionPrerequisites, erosion, emission, transportPrerequisites and/or transportCalcOrder beforehand.

# S4 method for RPhosFate
subsequentRun(
  x,
  substance = "PP",
  erosionPrerequisites = FALSE,
  erosion = FALSE,
  emission = FALSE,
  transportPrerequisites = FALSE,
  transportCalcOrder = FALSE
)

Arguments

x

An S4 RPhosFate river catchment object.

substance

A character string specifying the substance to calculate.

erosionPrerequisites

A logical scalar specifying if erosionPrerequisites is called.

erosion

A logical scalar specifying if erosion is called.

emission

A logical scalar specifying if emission is called. It is never called with substance = "SS" though.

transportPrerequisites

A logical scalar specifying if transportPrerequisites is called.

transportCalcOrder

A logical scalar specifying if transportCalcOrder is called.

Value

An S4 RPhosFate river catchment object and side effects in the form of raster files.

See also

Examples

if (FALSE) {
# temporary demonstration project copy
cv_dir <- demoProject()
# load temporary demonstration project
x <- RPhosFate(
  cv_dir = cv_dir,
  ls_ini = TRUE
)
# presupposed method call
x <- firstRun(x, "SS")

x <- subsequentRun(x, "PP")
}